Move the policies.h file to compute
